% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCoinApp.R
\name{runCoinApp}
\alias{runCoinApp}
\title{Launch the Coin Shiny App}
\usage{
runCoinApp()
}
\value{
Launches a Shiny application.
}
\description{
Opens the interactive Shiny application included in the
**Lab2KorbDSA5403** package. The app provides dynamic visualization
of the discrete‑theta Bayesian coin model, including prior,
likelihood, posterior, and credible intervals.
}
\examples{
\dontrun{
  runCoinApp()
}
}
